feat: add argument for cleaning first, stop making it a default

+++ fixLocalPackages.js
@@ -5,18 +5,27 @@ const { join } = require("path");
 const { useYarn } = require("yarnif");
 const rimraf = require("rimraf");
 const { existsSync } = require("fs");
-if (useYarn()) {
-  const { stdout } = spawnSync("yarn", ["cache", "dir"], { encoding: "utf-8" });
-  const cacheDir = stdout.toString().trim();
-  if (cacheDir) {
-    const tmp = join(cacheDir, ".tmp");
-    if (existsSync(tmp)) {
-      rimraf.sync(tmp);
-      console.log("Removed all yarn cache contents from ", tmp);
-    } else console.log("No temp dir at path", tmp);
-  } else {
-    console.error("No results from yarn cache dir");
+const commander = require("commander");
+commander.arguments(["[path]"]);
+commander.option("-c --clean", "Clean the temp directory first");
+commander.parseAsync(process.argv).then(() => {
+  if (commander.clean) {
+    if (useYarn()) {
+      const { stdout } = spawnSync("yarn", ["cache", "dir"], {
+        encoding: "utf-8",
+      });
+      const cacheDir = stdout.toString().trim();
+      if (cacheDir) {
+        const tmp = join(cacheDir, ".tmp");
+        if (existsSync(tmp)) {
+          rimraf.sync(tmp);
+          console.log("Removed all yarn cache contents from ", tmp);
+        } else console.log("No temp dir at path", tmp);
+      } else {
+        console.error("No results from yarn cache dir");
+      }
+    }
   }
-}
-const out = fixLocalPackages(process.argv[2] || process.cwd());
-console.log("All done");
+  const out = fixLocalPackages(commander.args[0] || process.cwd());
+  console.log("All done");
+});
